+++ tests/functional_tests/test_general_formatting.py
@@ -54,6 +54,7 @@ from tests.util import check_formatting, read_test_data
         "unformatted/213_gitlab_fct_sales_funnel_target.sql",
         "unformatted/300_jinjafmt.sql",
         "unformatted/400_create_fn_and_select.sql",
+        "unformatted/401_explain_select.sql",
     ],
 )
 def test_formatting(p: str) -> None: