+++ tests/unit_tests/test_dialect.py
@@ -191,6 +191,7 @@ class TestAllDialects:
             ("main", "set_operator", "minus"),
             ("main", "set_operator", "except"),
             ("main", "bq_typed_array", "array<INT64>"),
+            ("main", "nonreserved_keyword", "explain"),
             ("main", "unsupported_ddl", "create table"),
             ("main", "unsupported_ddl", "select\ninto"),
             ("main", "unsupported_ddl", "insert"),